requirements.txt
asgiref==3.3.4
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cryptography==3.4.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.2
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
gTTS-token==1.1.3
idna==2.8
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
